Fix category assignment.
1. Category label should not be quoted.
2. WordPress's concept of an uncategorized entry involves an actual category named "Uncategorized" (I have no idea WTF.)
If an MT entry has no category assigned and is passed through that way, WP will /display/ it as "Uncategorized" but in actuality the category is null.
This caused some problem I can't recall at the moment.
So, we now first test for category assignment, and explicitly set to "Uncategorized" if there is none.